[ESLint] Fix AdminScreen
