+++ .eslintrc.json
@@ -23,6 +23,7 @@
         "react"
     ],
     "rules": {
+        // Custom Rules
         "indent": "off",
         "react/jsx-indent": ["error", 4],
         "react/jsx-indent-props": ["error", 4],
@@ -30,13 +31,13 @@
         "comma-dangle": ["error", "never"],
 
         // TODO: Remove these temporary rules
-        "jsx-a11y/anchor-is-valid": "off",
         "react/no-unused-state": "off",
         "react/destructuring-assignment": "off",
         "max-len": "off",
-
+        "react/prop-types": "off",
 
         // TODO: Re-enable these accessibility rules
+        "jsx-a11y/anchor-is-valid": "off",
         "jsx-a11y/no-static-element-interactions": "off",
         "jsx-a11y/click-events-have-key-events": "off",
         "jsx-a11y/interactive-supports-focus": "off",
